refactor(frontend): split api.ts into api/ module (Commit 1)
Split the 417-line api.ts into 6 focused files under frontend/lib/api/:
client.ts, runtime.ts, conversations.ts, routes.ts, feedback.ts, index.ts.

The barrel index.ts re-exports everything so all existing imports remain
unchanged. Adds fetchPopularBangumi() in client.ts for the upcoming
WelcomeScreen. Adds tests/api-split.test.ts to verify all exports resolve.
